Collapse util files
